Clean up and renamed files.
