+++ .github/workflows/build-proto-cf-handler.yml
@@ -47,7 +47,7 @@ jobs:
         with:
           func: generic-cloudevent-handler
           module: testrunner
-          entry_point: bio.terra.cloudfunctions.streaming.GenericCFEventHandler
+          entry_point: bio.terra.cloudfunctions.proto.ProtoFunc
           runtime_memory: 512MB
           bucket: testrunner-results
           trigger: google.storage.object.finalize
